use time to flush lines instead of volume to time conversion

+++ State.h
@@ -22,6 +22,6 @@ bool state_isStepComplete(void);
 
 void state_copyStateData(system_t*);
 
-uint16_t state_remainingMillis(void);
+uint32_t state_remainingMillis(void);
 
 void state_checkSkipRinse(uint8_t);
